+++ scriptconfig/dict_like.py
@@ -27,9 +27,38 @@ class DictLike:
     ``__delitem__``, ``__setitem__``, ``update``,
 
     And perhaps: ``copy``,
+
+
+    Example:
+        from scriptconfig.dict_like import DictLike
+        class DuckDict(DictLike):
+            def __init__(self, _data=None):
+                if _data is None:
+                    _data = {}
+                self._data = _data
+
+            def getitem(self, key):
+                return self._data[key]
+
+            def keys(self):
+                return self._data.keys()
+
+        self = DuckDict({1: 2, 3: 4})
+        print(f'self._data={self._data}')
+        cast = dict(self)
+        print(f'cast={cast}')
+        print(f'self={self}')
+
     """
 
     def getitem(self, key):
+        """
+        Args:
+            key (Any): the key
+
+        Returns:
+            Any : the associated value
+        """
         raise NotImplementedError('abstract getitem function')
 
     def setitem(self, key, value):
@@ -39,6 +68,10 @@ class DictLike:
         raise NotImplementedError('abstract delitem function')
 
     def keys(self):
+        """
+        Yields:
+            str:
+        """
         raise NotImplementedError('abstract keys function')
 
     def __repr__(self):
@@ -66,15 +99,16 @@ class DictLike:
         return self.setitem(key, value)
 
     def items(self):
-        return self.iteritems()
+        return ((key, self[key]) for key in self.keys())
 
     def values(self):
-        return self.itervalues()
+        return (self[key] for key in self.keys())
 
     def copy(self):
         return dict(self.items())
 
     def asdict(self):
+        # Alias for to_dict
         return dict(self.items())
 
     def to_dict(self):
@@ -86,12 +120,21 @@ class DictLike:
             self[k] = v
 
     def iteritems(self):
+        import ubelt as ub
+        ub.schedule_deprecation(
+            'scriptconfig', 'iteritems', 'use items instead')
         return ((key, self[key]) for key in self.keys())
 
     def itervalues(self):
+        import ubelt as ub
+        ub.schedule_deprecation(
+            'scriptconfig', 'itervalues', 'use items instead')
         return (self[key] for key in self.keys())
 
     def iterkeys(self):
+        import ubelt as ub
+        ub.schedule_deprecation(
+            'scriptconfig', 'iterkeys', 'use items instead')
         return (key for key in self.keys())
 
     def get(self, key, default=None):